Refactor: Allow experimental job traces   - Pcon fields are inline doubles—no per-job vectors, pointers, or allocations.   - Standard records/resource samples retain their original sizes.   - Added policy-based Trace/PconTrace and Simulation/PconSimulation.   - Pcon totals are tracked through job start/finish and written to resource traces.   - Removed scheduler dependence on concrete trace storage.   - No DR_EVT_PCON_EXPERIMENT macro.   - Cleaned stale scheduler documentatio

+++ src/trace/parse_utils.cpp
@@ -115,7 +115,7 @@ void set_by(unsigned &v, const std::string &str) {
 
 void set_by(double &v, const std::string &str) {
   size_t pos;
-  v = stof(str, &pos);
+  v = stod(str, &pos);
   if (pos != str.size()) {
     throw std::invalid_argument{"Failed to parse a double precision number! " +
                                 str};
